requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
coverage==5.3
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==5.8.0
flake8==3.8.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
